requirements.txt
websockets>=12.0
aiohttp>=3.9.0
python-dotenv>=1.0.0
pandas>=2.1.0
requests>=2.31.0
py-clob-client>=0.19.0
web3>=6.0.0